% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readSomaLogic.R
\name{isPass}
\alias{isPass}
\title{Is the value a pass}
\usage{
isPass(x)
}
\arguments{
\item{x}{A character vector or factor.}
}
\value{
A logical vector, the same length as the input, which is \code{TRUE}
whenever the input is the string "PASS".  Missing values return \code{FALSE}.
}
\description{
Checks if a string is "PASS".
}

