% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/soma2eset.R
\name{soma2eset}
\alias{soma2eset}
\title{Convert WideSomaLogicData into ExpressionSet}
\usage{
soma2eset(somaObj, sampleDF = data.frame(sampleId = somaObj$SampleId,
  row.names = somaObj$SampleId), log2Transform = TRUE)
}
\arguments{
\item{somaObj}{WideSomaLogicData object}

\item{sampleDF}{dataframe with sample annotations}

\item{log2Transform}{whether to log2 transform intensities or not}
}
\value{
ExpressionSet object
}
\description{
Convert WideSomaLogicData into ExpressionSet
}
\author{
Aditya Bhagwat
}

