% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readSomaLogic.R
\name{[.WideSomaLogicData}
\alias{[.WideSomaLogicData}
\title{Indexing for WideSomaLogicData objects}
\usage{
\method{[}{WideSomaLogicData}(x, ...)
}
\arguments{
\item{x}{A \code{WideSomaLogicData} object.}

\item{...}{Passed to \code{[.data.table}.}
}
\value{
A \code{WideSomaLogicData} object.
}
\description{
Wrapper to \code{[.data.table}, ensuring that the \code{SequenceInfo},
\code{Metadata} and \code{Checksum} attributes are preserved.
}
\seealso{
\code{\link[data.table]{data.table}}
}

